requirements.txt
python-dotenv
discord.py
mcstatus
matplotlib